# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserInfo"
      required:
        - token
        - user
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, operator, student]
        institutionId:
          type: integer
          nullable: true
      required:
        - id
        - email
        - name
        - role
    Institution:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - isActive
        - createdAt
        - updatedAt
    CreateInstitutionRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
      required:
        - name
    UpdateInstitutionRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        isActive:
          type: boolean
    DashboardStats:
      type: object
      properties:
        totalInstitutions:
          type: integer
        totalStudents:
          type: integer
        totalOperators:
          type: integer
        activeReaders:
          type: integer
      required:
        - totalInstitutions
        - totalStudents
        - totalOperators
        - activeReaders
    CreateOperatorRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password:
          type: string
        institutionId:
          type: integer
      required:
        - email
        - name
        - password
        - institutionId
    OperatorInfo:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        institutionId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        institution:
          $ref: "#/components/schemas/Institution"
      required:
        - id
        - email
        - name
        - role
...
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
          nullable: true
        isbn:
          type: string
          nullable: true
        language:
          type: string
        level:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - language
        - isActive
        - createdAt
        - updatedAt
...
    BookAssignment:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        studentId:
          type: integer
        institutionId:
          type: integer
        assignedBy:
          type: integer
          nullable: true
        assignedAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
        book:
          $ref: "#/components/schemas/Book"
        institution:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
      required:
        - id
        - bookId
        - studentId
        - institutionId
        - assignedAt
        - isActive
...
    StudentTokenRequest:
      type: object
      properties:
        idToken:
          type: string
          description: OAuth ID token (Google) or access token (Microsoft)
        provider:
          type: string
          enum: [google, microsoft]
          description: OAuth provider
      required:
        - idToken
        - provider
    AdminActivity:
      type: object
      properties:
        recentAssignments:
          type: array
          items:
            $ref: "#/components/schemas/BookAssignment"
        recentProgress:
          type: array
          items:
            $ref: "#/components/schemas/ReadingProgress"
        recentUsers:
          type: array
          items:
            $ref: "#/components/schemas/UserInfo"
      required:
        - recentAssignments
        - recentProgress
        - recentUsers
    ReadingProgress:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        bookId:
          type: integer
        assignmentId:
          type: integer
          nullable: true
        pagesRead:
          type: integer
        totalPages:
          type: integer
          nullable: true
        completionPercent:
          type: integer
        lastReadAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        student:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            email:
              type: string
        book:
          type: object
          properties:
            id:
              type: integer
            title:
              type: string
            author:
              type: string
              nullable: true
      required:
        - id
        - studentId
        - bookId
        - pagesRead
        - completionPercent
        - startedAt
        - createdAt
        - updatedAt
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/student-token:
    post:
      operationId: studentTokenExchange
      tags: [auth]
      summary: Exchange OAuth credentials for a student JWT token
      description: Creates or finds a student user by email and returns a JWT token. Used after OAuth login to bridge into the dashboard auth system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentTokenRequest"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get global dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/institutions:
    get:
      operationId: listInstitutions
      tags: [admin]
      summary: List all institutions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of institutions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Institution"
    post:
      operationId: createInstitution
      tags: [admin]
      summary: Create a new institution
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstitutionRequest"
      responses:
        "201":
          description: Institution created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Institution"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/institutions/{id}:
    get:
      operationId: getInstitution
      tags: [admin]
      summary: Get institution by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Institution details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Institution"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateInstitution
      tags: [admin]
      summary: Update an institution
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInstitutionRequest"
      responses:
        "200":
          description: Institution updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Institution"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteInstitution
      tags: [admin]
      summary: Delete an institution
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Institution deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/operators:
    get:
      operationId: listOperators
      tags: [admin]
      summary: List all operators
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of operators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OperatorInfo"
    post:
      operationId: createOperator
      tags: [admin]
      summary: Create an operator account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOperatorRequest"
      responses:
        "201":
          description: Operator created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/activity:
    get:
      operationId: getAdminActivity
      tags: [admin]
      summary: View all recent activity across the platform
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
          description: Maximum number of items per category (max 200)
      responses:
        "200":
          description: Recent activity feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminActivity"
...