# lib/api-zod/src/generated/types/adminActivity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookAssignment } from "./bookAssignment";
import type { ReadingProgress } from "./readingProgress";
import type { UserInfo } from "./userInfo";

export interface AdminActivity {
  recentAssignments: BookAssignment[];
  recentProgress: ReadingProgress[];
  recentUsers: UserInfo[];
}